requirements.txt
# pip install -r requirements.txt
django
psycopg2-binary
django-unfold
celery
redis
django-celery-beat
pyyaml
requests